Move constraint check to IliValidatorLogParser
